Create separate package for screens and move Composables into it
